tests/baseline/protocol-1.2/inspector/sources/debugger/source-frame-inline-breakpoint-decorations-expected.txt

Checks that JavaScriptSourceFrame show inline breakpoints correctly


Running: testAddRemoveBreakpoint
Setting breakpoint
breakpoint at 4
Toggle breakpoint
breakpoint at 4

Running: testAddRemoveBreakpointInLineWithOneLocation
Setting breakpoint
breakpoint at 4
breakpoint at 5
Toggle breakpoint
breakpoint at 4

Running: clickByInlineBreakpoint
Setting breakpoint
breakpoint at 4
Click by second breakpoint
Could not click on Javascript breakpoint - lineNumber: 3, index: 1

Running: toggleBreakpointInAnotherLineWontRemoveExisting
Setting breakpoint in line 4
Click by first breakpoint
Could not click on Javascript breakpoint - lineNumber: 3, index: 0
